삼성코테 주사위굴리기.cpp

+++ 삼성코테/시험감독.cpp
@@ -4,7 +4,7 @@ using namespace std;
 
 int N, B, C;
 vector<int> v;
-int gamdok = 0;
+long long gamdok = 0;
 
 void input() {
     cin >> N;
